Consulta Agregado

+++ src/main/java/com/omy/gmedica/dao/IConsultaDao.java
@@ -11,7 +11,7 @@ import com.omy.gmedica.model.Medico;
 import com.omy.gmedica.model.Paciente;
 
 @Repository
-public interface IConsultaDao extends JpaRepository<Consulta, Integer> {
+public interface IConsultaDao extends JpaRepository<Consulta, Long> {
 	/*List<Consulta> findByPacienteId(int idPaciente);
 
 	List<Consulta> findByEspecialidadId(int idEspecialidad);
